+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@ Version 1.2.1 (TBD): Not yet released
 [@GooberRF](https://github.com/GooberRF)
 - Change default spawn delay for KOTH mode from 5 seconds to 2 seconds
 - Drop support for Direct3D 11 renderer on GPUs without D3D level 10_0 (or higher) support
+- Improve FPS/ping HUD display
+- Change prefix for `showfps` and `showping` console commands to `ui`
+- Add `ui_fpsavg` command set FPS counter averaging window
 
 [@is-this-c](https://github.com/is-this-c)
 - Improve freelook spectate logic after level transition